+++ zim/gui/preferencesdialog.py
@@ -500,7 +500,8 @@ class KeyBindingsTab(Gtk.VBox):
 	def __init__(self, dialog):
 		GObject.GObject.__init__(self)
 		help = _(
-			'Key bindings can be changed by selecting the action in the list and pressing the new key binding.\n'
+			'Key bindings can be changed by clicking on a field with a key combination\n'
+			'in the list and then press the new key binding.\n'
 			'To disable a keybinding, select it in the list and use <tt>&lt;Backspace&gt;</tt>.'
 		) # T: help text in preferences dialog for modifying keybindings
 		self.add(ScrolledWindow(KeyBindingTreeView()))